removed a number of warnings

+++ sejda-sambox/src/main/java/org/sejda/impl/sambox/component/image/ImagesToPdfDocumentConverter.java
@@ -32,6 +32,7 @@ import org.sejda.impl.sambox.component.PDDocumentHandler;
 import org.sejda.impl.sambox.component.PageImageWriter;
 import org.sejda.model.PageOrientation;
 import org.sejda.model.PageSize;
+import org.sejda.model.encryption.EncryptionAtRestPolicy;
 import org.sejda.model.exception.TaskException;
 import org.sejda.model.exception.TaskIOException;
 import org.sejda.model.input.ImageMergeInput;
@@ -39,7 +40,6 @@ import org.sejda.model.input.MergeInput;
 import org.sejda.model.input.PdfFileSource;
 import org.sejda.model.input.PdfMergeInput;
 import org.sejda.model.input.Source;
-import org.sejda.model.encryption.EncryptionAtRestPolicy;
 import org.sejda.model.parameter.BaseMergeParameters;
 import org.sejda.model.task.TaskExecutionContext;
 import org.sejda.sambox.pdmodel.PDPage;
@@ -109,7 +109,7 @@ public class ImagesToPdfDocumentConverter {
                     int targetHeight = (int) mediaBox.getHeight();
                     LOG.debug("Scaling image down to fit by height {} vs {}", height, targetHeight);
 
-                    float ratio = (float) height / targetHeight;
+                    float ratio = height / targetHeight;
                     height = targetHeight;
                     width = Math.round(width / ratio);
                 }
